# doc/api.yml
openapi: 3.0.3
info:
  version: 'latest'
  title: Image Annotation API
servers:
  - url: 'https://annotator-api.imagery.bpartners.app'
security:
  - BearerAuth: [ ]
  - apiKeyHeader: [ ]
paths:
...
  '/jobs':
    get:
      tags:
        - Jobs
        - Admin
      summary: Get all jobs
      security: [ apiKeyHeader: [ ] ]
      operationId: getJobs
      parameters:
        - in: query
          name: page
          schema:
            $ref: "#/components/schemas/Page"
        - in: query
          name: pageSize
          schema:
            $ref: "#/components/schemas/PageSize"
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/JobStatus"
        - in: query
          name: name
          description: filters jobs by name
          schema:
            type: string
        - in: query
          name: type
          description: filters by job type
          schema:
            $ref: "#/components/schemas/JobType"
      responses:
        '200':
          description: List of all jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Job'
        '500':
          $ref: '#/components/responses/500'
...
  '/annotated-jobs/{jobId}':
    put:
      tags:
        - Jobs
        - Admin
        - AnnotatedJobs
      security:
        - apiKeyHeader: [ ]
      summary: create or update an annotated job
      operationId: crupdateAnnotatedJob
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrupdateAnnotatedJob"
      responses:
        '200':
          description: the job created with the givenId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        '500':
          $ref: '#/components/responses/500'
  '/jobs/{jobId}':
    get:
      tags:
        - Jobs
        - Admin
      security: [ apiKeyHeader: [ ] ]
      summary: get a job specified by its id
      operationId: getJob
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: the job with the given id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    put:
      tags:
        - Jobs
        - Admin
      security: [ apiKeyHeader: [ ] ]
      summary: Create or update a job
      operationId: saveJob
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrupdateJob'
      responses:
        '200':
          description: The created or updated job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '500':
          $ref: '#/components/responses/500'
  '/jobs/{jobId}/annotationStatistics':
    get:
      tags:
        - Jobs
      operationId: getJobLatestAnnotationStatistics
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The created or updated job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnotationStatistics'
        '500':
          $ref: '#/components/responses/500'
  '/jobs/{jobId}/export':
      get:
        tags:
          - Jobs
          - Admin
        operationId: exportJob
        description: exports a job to COCO or VGG format
        parameters:
          - name: jobId
            required: true
            in: path
            content:
              application/json:
                schema:
                  type: string
          - name: format
            required: true
            in: query
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ExportFormat"
          - name: emailCC
            description: one email address which will receive the exported job other than the ownerEmail, will default to none
            required: false
            in: query
            content:
              application/json:
                schema:
                  type: string
        responses:
          "200":
            description: OK response for mail sending initiation
            content:
              text/plain:
                schema:
                  type: string
                  example: "OK"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    apiKeyHeader:
      type: apiKey
      in: header
      name: x-api-key
  responses:
...
    '500':
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerException'
  schemas:
...
    Page:
      type: integer
      description: Start at `1`
    PageSize:
      type: integer
      description: Cannot exceed `500`
...
    CrupdateJob:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/JobType"
        id:
          type: string
        name:
          type: string
        bucketName:
          type: string
        folderPath:
          type: string
        ownerEmail:
          type: string
          format: email
        imagesHeight:
          type: integer
        imagesWidth:
          type: integer
        status:
          $ref: '#/components/schemas/JobStatus'
        teamId:
          type: string
        labels:
          type: array
          items:
            $ref: "#/components/schemas/Label"
    JobType:
      type: string
      enum:
        - LABELLING
        - REVIEWING
    CrupdateAnnotatedJob:
      type: object
      allOf:
        - $ref: "#/components/schemas/CrupdateJob"
      properties:
        annotatedTasks:
          type: array
          items:
            $ref: "#/components/schemas/AnnotatedTask"
    AnnotatedTask:
      type: object
      properties:
        id:
          type: string
        annotatorId:
          type: string
        filename:
          type: string
          pattern: "^[^/]"
        annotationBatch:
          $ref: "#/components/schemas/AnnotationBatch"
    Job:
      type: object
      allOf:
        - $ref: "#/components/schemas/CrupdateJob"
      properties:
        taskStatistics:
          $ref: "#/components/schemas/TaskStatistics"
    JobStatus:
      type: string
      enum:
        - PENDING
        - READY
        - STARTED
        - TO_REVIEW
        - TO_CORRECT
        - FAILED
        - COMPLETED
...
    TaskStatistics:
      type: object
      properties:
        remainingTasksForUserId:
          type: integer
          format: int64
        remainingTasks:
          type: integer
          format: int64
        completedTasksByUserId:
          type: integer
          format: int64
        totalTasks:
          type: integer
          format: int64
    AnnotationNumberPerLabel:
      type: object
      properties:
        labelName:
          type: string
        numberOfAnnotations:
          type: integer
          format: int64
    AnnotationStatistics:
      type: array
      items:
        $ref: "#/components/schemas/AnnotationNumberPerLabel"
...
    Annotation:
      type: object
      properties:
        id:
          type: string
        taskId:
          type: string
        userId:
          type: string
        label:
          $ref: '#/components/schemas/Label'
        polygon:
          $ref: '#/components/schemas/Polygon'
    AnnotationBatch:
      type: object
      properties:
        id:
          type: string
        annotations:
          type: array
          items:
            $ref: "#/components/schemas/Annotation"
        creationDatetime:
          description: ignored in requestBody
          type: string
          format: date-time
    Polygon:
      type: object
      properties:
        points:
          type: array
          items:
            $ref: '#/components/schemas/Point'
    Point:
      type: object
      properties:
        x:
          type: number
          format: double
        y:
          type: number
          format: double
    Label:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
          description: The color to assign to the label into hexadecimal form
          example: "#40D399"
...
    ExportFormat:
      type: string
      enum:
        - COCO
        - VGG
    Exception:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
    InternalServerException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: InternalServerException
        message: Unexpected error
...